@@ -1,10 +1,11 @@
 using Core.Entity;
+using System.Linq.Expressions;
 
 namespace Application.Interface;
 
 public interface ICategoryService
 {
-    IEnumerable<Category> GetAllCategories();
+    IEnumerable<Category> GetAllCategories(Expression<Func<Category, bool>> expression = null);
 
     Task<bool> Insert(Category category);
 
