@@ -1,12 +1,13 @@
 using Core.Entity;
+using System.Linq.Expressions;
 
 namespace Application.Interface;
 
 public interface IColorService
 {
-    IEnumerable<Color> GetAllColors();
+    IEnumerable<Color> GetAllColors(Expression<Func<Color, bool>> expression = null);
 
-    Task<sbyte> Save(Color color);
+    Task<bool> Save(Color color);
 
     Task<bool> Lock(string code);
 
