@@ -1,11 +1,12 @@
 using Core.Entity;
 using Microsoft.AspNetCore.Http;
+using System.Linq.Expressions;
 
 namespace Application.Interface;
 
 public interface ISportService
 {
-    IEnumerable<Sport> GetAllSports();
+    IEnumerable<Sport> GetAllSports(Expression<Func<Sport, bool>> expression = null);
 
     Task<bool> Insert(Sport sport);
 
