@@ -2,6 +2,7 @@ using Application.Interface;
 using Core.DTO;
 using Core.Entity;
 using Core.Interface;
+using System.Linq.Expressions;
 
 namespace Application.Service;
 
@@ -23,7 +24,7 @@ public class CartService : ICartService
 	{
 		CartDetail cart = new CartDetail
         {
-            UserId = _unitOfWork.Users.GetUserIdByUsername(cd.Username),
+            UserId = await _unitOfWork.Users.GetUserIdByUsername(cd.Username),
             Sku = cd.Sku,
             Price = cd.Price,
             Quantity = cd.Quantity
@@ -37,7 +38,7 @@ public class CartService : ICartService
 	{
 		CartDetail cart = new CartDetail
         {
-            UserId = _unitOfWork.Users.GetUserIdByUsername(cd.Username),
+            UserId = await _unitOfWork.Users.GetUserIdByUsername(cd.Username),
             Sku = cd.Sku,
             Price = cd.Price,
             Quantity = cd.Quantity
@@ -49,13 +50,13 @@ public class CartService : ICartService
 
     public async Task<bool> Delete(string username, string sku)
 	{
-		_unitOfWork.Carts.Delete(_unitOfWork.Users.GetUserIdByUsername(username), sku);
+		_unitOfWork.Carts.Delete(await _unitOfWork.Users.GetUserIdByUsername(username), sku);
 		return await _unitOfWork.SaveChangesAsync();
 	}
 
     public async Task<bool> DeleteAll(string username)
 	{
-		_unitOfWork.Carts.DeleteAll(_unitOfWork.Users.GetUserIdByUsername(username));
+		_unitOfWork.Carts.DeleteAll(await _unitOfWork.Users.GetUserIdByUsername(username));
 		return await _unitOfWork.SaveChangesAsync();
 	}
 }
