@@ -12,7 +12,7 @@ public interface IBlogRepository : IInsert<BlogArticle>, IUpdate<BlogArticle>, I
 
     Task<BlogDetailReadDTO> GetByUrlAsync(string url);
 
-    IEnumerable<BlogArticle> GetAllBlogs();
+    IEnumerable<BlogArticle> GetAllBlogs(Expression<Func<BlogArticle, bool>> expression = null);
 
     BlogArticle GetBlogById(int id);
 }
